// src/shared/utils/injection.enum.ts
export enum InjectionEnum {
  AUTH_SERVICE = 'AUTH_SERVICE',
  PASSWORD_ENCRYPTOR = 'PASSWORD_ENCRYPTOR',
  USER_REPOSITORY = 'USER_REPOSITORY',
  USER_SERVICE = 'USER_SERVICE',
  TOKEN_MANAGER = 'TOKEN_MANAGER',
  MOVIE_SERVICE = 'MOVIE_SERVICE',
  MOVIE_REPOSITORY = 'MOVIE_REPOSITORY',
  ACTOR_SERVICE = 'ACTOR_SERVICE',
  GENDER_SERVICE = 'GENDER_SERVICE',
  DIRECTOR_SERVICE = 'DIRECTOR_SERVICE',
  ACTOR_REPOSITORY = 'ACTOR_REPOSITORY',
  GENDER_REPOSITORY = 'GENDER_REPOSITORY',
  DIRECTOR_REPOSITORY = 'DIRECTOR_REPOSITORY',
  ROLE_SERVICE = 'ROLE_SERVICE',
  ROLE_REPOSITORY = 'ROLE_REPOSITORY',
}